Rdkafka API uses int for timeout

+++ src/KafkaW/BrokerSettings.h
@@ -15,7 +15,7 @@ public:
   BrokerSettings();
   void apply(rd_kafka_conf_t *RdKafkaConfiguration);
   std::string Address;
-  size_t PollTimeoutMS = 100;
+  int PollTimeoutMS = 100;
   std::map<std::string, int> ConfigurationIntegers;
   std::map<std::string, std::string> ConfigurationStrings;
 };
